Create snake.cpp (#5)
Code I currently have. Snake moves on key presses. A segment body to start. Food is on an algorithm to pop up in different places, but we probably want a vector of rigged locations as suggested in the chat. The snake can die if it hits the wall or touches itself.